WSP 62 Refactoring Phase 2: Extract training system from main.py
Problem: main.py still in WSP 62 critical window (1748 lines after Phase 1)
Solution: Extract massive training system functions to dedicated module

Changes:
- main.py: 1748 → 1316 lines (-432 lines, 24.7% reduction)
- Created training_system module with 2 major functions
- WSP 62 COMPLIANT: 184 lines BELOW critical window threshold

Extracted Functions:
1. run_training_system() → modules/ai_intelligence/training_system/scripts/launch.py
   - 321 lines: Complete Qwen/Gemma training submenu
   - 9 training options: batch training, UTF-8 hygiene, pattern recall, routing tests
   - WRE pattern implementation (WSP 46)

2. execute_training_command() → modules/ai_intelligence/training_system/scripts/training_commands.py
   - 115 lines: Headless training command executor for 0102
   - Commands: utf8_scan, utf8_summary, utf8_fix, batch
   - JSON output support for automation

WSP Compliance:
- ✅ WSP 62: NOW COMPLIANT - 1316 lines (well below 1500 threshold)
- ✅ WSP 3: Proper ai_intelligence domain placement
- ✅ WSP 49: training_system/scripts/ structure
- ✅ WSP 22: Complete traceable narrative

Total Progress (Phase 1 + Phase 2):
- Original: 1951 lines (WSP 62 violation)
- After Phase 1: 1748 lines (-203)
- After Phase 2: 1316 lines (-432)
- **Total Reduction: -635 lines (32.5%)**
- **Status: WSP 62 COMPLIANT**

Remaining Functions (optional future work):
- run_utf8_hygiene_scan(): 130 lines
- check_instance_status(): 110 lines
- view_git_post_history(): 83 lines
- summarize_utf8_findings(): 66 lines
- generate_x_content(): 42 lines
- git_push_and_post(): 22 lines

If extracted, main.py would be ~863 lines (even lighter weight)
